Move FilePane into its own file. Also start adding action bar to FilePane.
